String function cleanup and lighting fixes.
Fixed lighting so items in inventories once again emit light.
Created some enum classes to make some optional arguments a bit easier to read.
Got rid of get_identifying_string_without_possessives entirely, making possessives an optional argument to get_identifying_string.
Turned off some loggers that I don't need on right now.

+++ src/MapTile.cpp
@@ -232,9 +232,9 @@ void MapTile::add_light_influence(ThingId source,
         float light_factor = (1.0f - dist_factor);
         float wall_factor = Direction::calculate_light_factor(influence.coords, get_coords(), d);
 
-        addColor.r = static_cast<sf::Uint8>(light_color.r * wall_factor * light_factor);
-        addColor.g = static_cast<sf::Uint8>(light_color.g * wall_factor * light_factor);
-        addColor.b = static_cast<sf::Uint8>(light_color.b * wall_factor * light_factor);
+        addColor.r = static_cast<sf::Uint8>(static_cast<float>(light_color.r) * wall_factor * light_factor);
+        addColor.g = static_cast<sf::Uint8>(static_cast<float>(light_color.g) * wall_factor * light_factor);
+        addColor.b = static_cast<sf::Uint8>(static_cast<float>(light_color.b) * wall_factor * light_factor);
         addColor.a = 255;
 
         unsigned int index = d.get_map_index();
